requirements/_base.txt
# core
djangorestframework
django_taggit

# needed by the doc views
coreapi
coreschema
uritemplate
